Created reusable accordian component
